Finished Cleaning and Normalizing data
Tried using some preprocessing stuff from sklearn at the recommendation of a stack exchange post, but I wanted to keep my dataframe unperturbed, and something was getting lost in translation. (Also, I accidentally turned a cell of code into markdown that was supposed to fill in NaN values with average values).

The solution was just a function that normalized the dataframe the old fashioned way. The price column was excluded from the normalization, since it is our target.